# Group offloading with cuda stream prefetching